clase04 completada
