+++ manipulando-dom-con-js/script.js
@@ -1,5 +1,7 @@
-( () => {
-    const btn = document.querySelector("[data-from-btn]")
+import checkComplete from './components/checkComplete.js'
+import deleteIcon from './components/deleteIcon.js'    
+
+const btn = document.querySelector("[data-from-btn]")
 
     const createTask = (evento) => {
         evento.preventDefault()
@@ -9,6 +11,7 @@
         const task = document.createElement('li')
         task.classList.add("card")
         input.value = "";
+        //* backticks
         const taskContent = document.createElement("div")
     
         const titleTask = document.createElement("span")
@@ -16,11 +19,11 @@
         titleTask.innerHTML = value
         taskContent.appendChild(checkComplete())
         taskContent.appendChild(titleTask)
-        const content = `
-        <i class="fas fa-trash-alt trashIcon icon"></i>`;
+    
        //* task.innerHTML = content;
         
        task.appendChild(taskContent)
+       task.appendChild(deleteIcon())
        list.appendChild(task);
         
     
@@ -29,19 +32,5 @@
 
     btn.addEventListener('click', createTask )
     
-    const checkComplete = () => {
-        const i = document.createElement("i");
-        i.classList.add('far', 'fa-check-square', 'icon')
-        i.addEventListener('click' ,completTask)
-        return i;
-    }
-    
-    //? Immendiately invoked function expression IIFE
-    const completTask = (event) => {
-        const element = event.target
-        element.classList.toggle('fas')
-        element.classList.toggle('completeIcon')
-        element.classList.toggle('far')
-    }
-})()
+
